Stab at fixing https://github.com/Microsoft/visualfsharp/issues/49
The provided types were only being merged into the compiled module at the
very end, so the sources, compiled before that point, could not find the
type definitions to extend. This patch causes the type provider type
definitions to be generated *before* the sources are compiled and added
to the TypeDefsBuilder like any other type. This way, the type definitions
can be extended.

To handle provided types that are nested in modules or other types, we
pass a filtered list of each nested provided type to each TypeDefsBuilder
to be added as the containing types are defined.

This patch still has a couple issues:

- The type checker is not seeing any new members added by the type
extension, so you get an UndefinedName exception if you try to reference
them (at least in the same compilation unit).

- This fix does not apply to fsi. Fsi doesn't follow the same steps fsc
takes to convert the provided types into ILTypeDefs. We could probably
refactor that code out of the StaticLinker in fsc.fs into code shared
with fsi.